refactored UserService so switch statement is encapsulated within a UserAction interface that creates different kinds of actions that can be easily extended with polymorphism as opposed to adding more code to UserService file
