+++ aws/image-processor-lambdas/src/main/java/lambdas/helpers/ApiGatewayResponseHelper.java
@@ -1,0 +1,14 @@
+package lambdas.helpers;
+
+import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyResponseEvent;
+
+import java.util.Collections;
+
+public interface ApiGatewayResponseHelper {
+    static APIGatewayProxyResponseEvent createApiGatewayResponse(int statusCode, String message) {
+        return new APIGatewayProxyResponseEvent()
+                .withStatusCode(statusCode)
+                .withHeaders(Collections.singletonMap("Content-Type", "text/plain"))
+                .withBody(message);
+    }
+}
