+++ aws/image-processor-lambdas/src/main/java/lambdas/helpers/DatabaseHelper.java
@@ -1,4 +1,4 @@
-package helpers;
+package lambdas.helpers;
 
 import com.zaxxer.hikari.HikariConfig;
 import com.zaxxer.hikari.HikariDataSource;
