+++ aws/image-processor-lambdas/src/main/java/lambdas/helpers/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package helpers;
+package lambdas.helpers;
 
 public class UserNotFoundException extends Exception {
     public UserNotFoundException() {
